dev-pecuaria-antigo/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==42.0.5
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-braces==1.15.0
django-crispy-forms==2.1
django-filter==24.1
django-import-export==3.3.7
django-nested-admin==4.0.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
idna==3.7
Markdown==3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-monkey-business==1.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
